requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
peewee==3.9.3
pyTelegramBotAPI==3.6.6
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
tinydb==3.13.0
urllib3==1.24.1
